requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
asyncpg>=0.29.0
pydantic>=2.7.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
binance-connector>=3.7.0
rich>=13.7.0
discord.py>=2.3.2
